Restrict workspace-level plugins to just socialcalc for now.
This isn't the full "ask the plugin what scope it wants" solution, but
socialcalc is the only plugin (currently) enabled/disabled at the workspace
level.  

We need to think more about the implications on how workspace-level
permissions are supposed to work when combined with account-level plugin
permissions, but don't have time to think about that right now.

git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@11332 593ce9bc-2028-0410-bb98-c566b643a948