Added helper methods to `ST::Bootstrap::OpenLDAP` to add/remove our LDAP configuration from the `ldap.yaml` file.
Implemented such that the configuration is saved out *automatically* when you
instantiate an OpenLDAP bootstrapper.  You no longer need to explicitly save
out the configuration *unless* you've gone in and changed it.

This helps push some things out of the unit tests that we do repeatedly, down
into the test harness.


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@11376 593ce9bc-2028-0410-bb98-c566b643a948