Don't leave dangling `slapd` processes between test runs.
When used under `st-bootstrap-openldap`, it *is* possible to have scenarios
where we've lost the configuration or log files that we were working with.

e.g.
    st-bootstrap-openldap start --daemonize
    ... run some tests
    ... nuke the build/test environment
    st-bootstrap-openldap (start|stop)

This scenario is actually common under our auto-builder if its doing "fresh"
builds.

We would end up with dangling processes because when the old bootstrapper is
killed off, we no longer have access to the AppConfig or log files that we'd
been started with.  That would cause `ST::LDAP::Config` to throw a fatal error
when we tried to remove ourselves from the LDAP configuration during our
cleanup.

This fatal error would then cause the rest of our `DESTROY { }` block to be
ignored/skipped and we'd never kill off the running OpenLDAP instance.  Thus,
a dangling OpenLDAP instance.


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@11390 593ce9bc-2028-0410-bb98-c566b643a948