Adaptive LDAP; if the User has their data changed/moved in LDAP, try to find them again.
In LDAP, we have *three* possible unique identifiers that we could use to
locate a User in the directory; dn, username, e-mail address.  Rather than
limiting ourselves to just _one_ of these when trying to find a "previously
known" User in LDAP, try them all.

Doing a wider search like this, we'll be more likely able to find a match
between a User in LDAP and their corresponding record in our DB.  This allows
us to update the information we have on that User locally, updating the User
as opposed to thinking he doesn't exist any more.


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@11464 593ce9bc-2028-0410-bb98-c566b643a948