`st-bootstrap-openldap` *must* be run as "root" on an Appliance; sudo automatically if the User forgot to do this.
This require of "must run as root on an Appliance" stems from how `ST::Hub`
and `ST::AppConfig` work; when `ST::Hub` is loaded it preloads all of the
plugins, some of which then try to read in some config.  They can't load the
config, though, as `ST::AppConfig` returns paths based on the starting User
being a *human* user instead of being the www-data/root User on an appliance.
Doing the sudo prevents errors like:

    Couldn't open /usr/share/share/fonts/config.yaml

Further, `st-bootstrap-openldap` will need to update the `socialtext.conf` and
`ldap.yaml` configuration files in `/etc/socialtext/`, which are owned by
"root".


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@11669 593ce9bc-2028-0410-bb98-c566b643a948