Let LDAP users enter their username on the "I forgot my password page", and show them the error *AFTER* we've determined that they're an LDAP user.
Many of our LDAP customers have mixed environments, where _some_ of the Users
are in fact still "Default" users, and may need to reset their passwords.
Previously, we'd show the error message right from the get go and these Users
would never be able to request that their password be reset.  Now, they have
the opportunity to go in there and do that.


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@11674 593ce9bc-2028-0410-bb98-c566b643a948