When forcably deleting Users from the DB, *first* re-assign their Workspaces *AND* Pages to the system-user.
Otherwise, we risk losing those Workspace/Page records entirely as we've now
broken referential integrity between the tables.

Fortunately, the *only* thing that forcably deletes User records is our test
suite.  We'll need this re-assignment, though, as we'll be making changes to
the DB momentarily that stop you dead in your tracks when doing this type of
DELETE if you haven't re-assigned Workspaces/Pages to some other User.


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@11721 593ce9bc-2028-0410-bb98-c566b643a948