Don't cascade User deletes across the "Workspace" and "page" tables, *RESTRICT* them.
Idea is to be more defensive in how we're building our DB schema.  If we allow
for User deletes to cascade, then deleting a User from the DB also causes all
of the Workspaces that they've created and all of the Pages that they either
created or were last to edit, to *also* be deleted from the DB.

BAD.  REALLY, REALLY BAD.

Instead, we'll switch this around so that we restrict the deletes in the DB
schema, and force you to stop and think about what you're actually doing.  If
you're using the API, it should just DTRT (as per r11721, which now re-assigns
before attempting the delete).  If you're operating _directly_ against the DB,
though, it now forces you to stop and think about what you're doing, lest you
accidentally delete a whole ton of data in the DB that you really don't want
to be deleting.


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@11722 593ce9bc-2028-0410-bb98-c566b643a948