Deprecate all code that "deletes users", moving it over into `Test::ST::User->delete_recklessly()`
We've known for ages that deleting a User is a bad (and dangerous) thing to be
doing.  So, *DON'T* expose any API methods that allow for this to be done.

This commit moves all of the functionality that _was_ implemented by the
various "delete the User/UserMetadata" methods out of our regular API and over
into a new `Test::Socialtext::User->delete_recklessly()` method.  This method
is named/labelled so that its far more obvious that (a) its for *test* code,
and (b) that its a dangerous operation to perform.

This commit also deprecates the need for `t/ST/User-delete-created-ty.t`, as
we don't expose the API calls that allow for this to happen any more.
Further, r11722 changes the underlying DB schema so that the nasty foreign key
cascading deletes this test was looking for aren't even possible any more
(which helps out in cases where you're playing directly against the DB and
aren't using our API).  Thus, this test becomes moot.  Thus, its removal.


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@11732 593ce9bc-2028-0410-bb98-c566b643a948