Internal restructuring, but functionally equivalent.
When reading pages and grabbing their data, move all the "push this page onto
the list of stuff to add" code outside of the eval block.  Makes it possible
for to short-circuit in the future (if needed), and also helps make it more
obvious which parts of this are likely to actually throw errors.


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@11737 593ce9bc-2028-0410-bb98-c566b643a948