Add LDAP mappings for profile fields as migration 33.
Uses the same set of defaults regardless of if the LDAP server is using AD or
OpenLDAP as it's driver.  We assume that most people will be using AD anyhow.
There's no good way to figure out on the fly what exactly the customer wants,
so we'll leave it up to them/CSMs to sort out the mapping.

Also update the ldap.yaml template with suggestions for AD and OpenLDAP.

git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@11940 593ce9bc-2028-0410-bb98-c566b643a948