Make `ST::Page::TablePopulator` actually support "recreate" as an *OPTION*.
Previously, it would *always* purge the Page table before re-populating it with
data.  Made the `--recreate` option in `st-populate-page-table` moot; even if
you didn't `--recreate` on the command line, `ST::Page::TablePopulator` would
still purge the Page table first.  Purging of the Page table would then cascade
through to other DB tables and cause a domino effect of things being wiped from
existence (e.g. Events).

This commit adds support for "recreate" as an option in
`ST::Page::TablePopulator`, which now allows for `st-populate-page-table` to be
run to just "fill in the missing blanks"; it'll add pages that were present on
disk but missing in the DB, but leave existing data in the DB alone.

Before committing, a search was done through our entire code base to find out
what made use of either `ST::Page::TablePopulator` or `st-populate-page-table`,
so that they could be updated to preserve their existing "recreate/purge"
behaviour.  Only *one* piece of code was found across our various SVN repos that
needed updating; `Socialtext::Workspace::Importer`.

During the above-mentioned search, it was found that the
`26-page-metadata-to-db` migration uses `ST::Page::TablePopulator`, *BUT* that
migration also has its own custom copy of `lib/` that it uses (so its *not*
using this updated version, ever).

To accompany this update, a pair of unit tests have been added to verify that
the "recreate" option works as intended (by purging the DB first if recreate=>1,
and by leaving the DB alone if recreate=>0).


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@11941 593ce9bc-2028-0410-bb98-c566b643a948