Resolved {bz: 1939}.
This bug is trigged because IE does not escape < and > in element attribute value. In other words,
doing this (let img_el be an <img> element):

    img_el.setAttribute("widget", "<p>123</p>\n123");

You get this element on Firefox:

    <img widget="&lt;p&gt;123&lt;/p&gt;
    123">

But on IE it's this:

    <IMG widget="<p>123</p>&#10;123">

( Also noticed that the \n is converted to only an &#10; (CR). )
Which in our dom-sanitizing process, badly converted to:

    <IMG widget="123<br class="p">&#10;123">

Then fed into the editor... *purr*

I am also pondering to put this attribute-escaping code in
Wikiwyg.create_element_with_attr routine. However that code is used
everywhere for quite a while. It should be carefully examined.