Record the IP address of the end-user in the Apache-perl access logs, not "127.0.0.1".
It has been discussed several times internally that various people would like to
see the *actual* IP address of the end-user in the Apache-perl access logs
instead of just "127.0.0.1" over and over.

The "Socialtext::Apache::ForwardedFor" module submitted in this commit will do
that, **IF** you comment out the line in ".../apache-perl/nlw.tt2".