Add a dev-script to dump users from configured LDAP servers.
This script is intended for diagnostics and isn't used by any code in our system.

Graham wrote the first version of this and I fixed it up to actually work
(needed 'base' and 'scope' params to the search).

Eventually we may want to move this into `bin` instead of just `dev-bin`.