Re-create signal events in this migration.
Signal events previously did not use the 'signal_id' column to refer back to
the signal table.  This migration deletes all of those old events and
re-creates them with information in the signal table.

Eventually, the signal table will contain (or have links regarding) other
content related to the signal.  The context.body field in the event should
contain the plain-text version of the signal.