When deleting Users recklessly, *don't* instantiate User objects (we may not be able to).
Some test cases end up in a state where we *can't* instantiate a User object,
so *don't* try to.  Instead, grab the UserMetadata directly in order to figure
out if the User was system created or not.

Do watch out, though, that some low-level Factory test suites create User
records _without_ creating a corresponding UserMetadata record.  When we
encounter this, *let* the User record get deleted; it should only be happening
for test/dummy Users, and you are *ONLY* supposed to be using this code inside
of test suites (that's why its called "Test::Socialtext::User").