Re-factor the `ST:U:F:ValidateAndCleanData()` method, splitting it up into a slew of smaller methods.
The refactor makes the function easier to read/follow, and moves each of the
individual validation/cleanup steps out into separate functions.  That in turn
makes each of those steps easier to read/follow, and we know that each of those
is doing "one thing and one thing only".

This refactor *also* allows us to eliminate some recursion that was present in
the previous implementation (where validating data would require that you
instantiate a homunculus object in order to test for duplicate data) (which
could fail spectacularly if we were in the middle of trying to instantiate a
homunculus object and needed to first validate said data).

Recursion was eliminated by adding a new `ST:U:D:F:lookup()` method which does
a lookup in the DB to see if the User data exists (and returning a proto-user
hash-ref if it does).  Room for further refactoring *is* present, as we've
leveraged the existing `GetHomunculus()` method to do the heavy lifting for us
here, and that _could_ be better refactored at a point in the future.

Functionally, this refactored version should behave *exactly* the same as it
did previously, but be easier to read and more flexible to work with and
reuse.