Handle LDAP User data validation errors properly.
Instead of trying to pass the data through to the DB, or letting a DBI error
message bubble itself up to the top, we now *actively* check for error
conditions and log the error appropriately.

With this error check in place, `st-refresh-ldap-users` *no longer* dies a
horrible death if it is unable to validate/cache the data for a User.  e.g. if
the User had no e-mail address in LDAP we don't choke and die any more but
instead display an error message and continue.