Turn off KeepAlive requests in the Apache/ModPerl back-end by default.
This is done by setting the special "nokeepalive" environment variable (which
is picked up by Apache, just like it does when you use the BrowserMatch
directive).  When done in this fashion, it is possible to have another part of
the back-end re-enable KeepAlive requests as needed, by setting
"nokeepalive=>undef".