small but important refacotring about wysiwyg mode loading.
The interal ready() event, when triggered, is almost meaningless to
IE. The reason for that, is the document are still empty (previous
set_inner_html hasn't totally finish yet).  Therefore, anything in there
depending on the document content (like grabbing all table.sort etc)
can't be working.

This commit uses a little piece of dom-ready checking code DOM.Ready
to test if the editor document is fully loaded.

Also it removes an extra calls to set_inner_html that was in the ready()
routine.