Eliminate the "--ports-start-at" option for `gen-config`.
The only place we've used it for something _other than_ "20000" is in
`Test::Live`, and by forcing that to now use "20000" just like everything else
the only advantage we lose is that "running live tests will occupy the same
ports you would normally use in your dev-env".  Considering how often we're
running the `t/live/` tests, I would not consider this a major setback.