Change the way fdefs sets up a basic dev-env, so that it can be done on a totally naked/clean environment.
Previously, fdefs would choke the very first time that you ran it against a
naked/clean environment (as with no ~/.nlw/ there when it starts,
`ST::AppConfig` thinks that its on an appliance, and it tries to add ceqlotron
files outside of your dev-env).

This update removes the `dev-bin/_fresh-dev-env` script and replaces it with a
new `dev-bin/make-test-fixture` script (which can be used more generically to
set up arbitrary test fixtures).

fdefs has been updated to use this new mechanism for setting up a dev-env, and
it is now possible to:

    rm -rf ~/.nlw/
    rm -rf t/tmp/
    dev-bin/fresh-dev-env-from-scratch

and have it work the *first* time.