Fix {bz: 2152} by rewriting the conversations query and adding an important index.
Ran the t/wikitest/rest/data_events_conversations.wiki tests to ensure that this didn't regress.

Cuts down the query time on staging for adina (who happens to be the user that's contributed to the most pages, one of the sore spots of the old query) from about 1.5 seconds down to about 80ms.