New `ST::Handler::Redirect`, to do simple HTTP redirects.
Although seemingly far too simple to warrant its own handler, it turns out to
be quite useful when relying on Apache to authenticate Users for us.  Send the
User off to an URL which requires Authentication, and when they provide
suitable credentials they're automatically redirected back to the page that
they wanted to be looking at in the first place.