Include POST params in log messages.
This will go a long way toward capturing any missing events in our
logs. We limit the length of any given param to 254 chars, the
maximum page length, so that we don't firehose the logs.

I am also adding some timing data to the AttachmentsUI so we
capture the amount of time it takes to upload files.