Adding a js-class submodule.
We decided to try doing submodules from the top level and then symlinking to then from whereever is needed.

We decided to namespace the submodules by github id so that people could easily switch to another fork of the module.
For instance, only ingy can commit to ingydotnet/js-class, but if gugod wants to make a fix he can fork to gugod/js-class and change the symlinks to point to his version. Then he would send ingy a pull request and ingy would incorporate gugod's changes and set all the pointers back, whenever he got to it.