Fix {bz: 2292}, allowing for DOS/Windows/Mac EOL.
Crux of the bug is that `Text::CSV_XS` would fail to parse the CSV file if it
contained non-native EOL characters.

We'll address this by auto-converting any CSV file that we've been given into
the native Unix-style EOL format *first*, and then letting `Text::CSV_XS` do
its thing.

Unit tests accompany the fix, to ensure that we don't regress on this in the
future.  I can confirm that the tests failed _before_ the fix, and that they
succeed _after_.