This is a mocked up contention warning.
This change will show a contention warning at the top of the page if someone else is editing it. THe message will fade away after 10 seconds. The time in the message is hardcoded.

Someone should be able to get this working properly quite easily.

Email ingy if you have questions.