Fix {bz: 2298} by using separate cookie for AIR/SD
Adobe AIR shares the Internet Explorer cookie store, which can cause all sorts
of conflict/problems for Users who use both on their desktop.  Logging in/out
of one automatically logs you in/out of the other, which isn't the most
intuitive thing for Users.

The behaviour we _want_ is that of "two separate applications", similarly to
how AIR behaves with Firefox or Safari; they're two separate apps and don't
share cookies or login credentials.

We fix this for Adobe AIR (and Socialtext Desktop) by sniffing the User-Agent
and then using a *different* HTTP cookie when we know that we've got an AIR
client.  Even if IE+AIR share the same cookie store (and may pass _both_
cookies along in their HTTP requests), we only ever care about or use *one* of
them (depending on what UA we've been given).

Yes, this feels klude-ish.  UA-sniffing always does.  The alternative, though,
is what several people have commented on as a crappy/non-intuitive user
experience.  Thanks Adobe. :|