Don't stomp on Admin WS, use our own.
Instead of using the Admin Workspace and creating a page in it with a name so
long that its going to cause problems, create a new dummy Test Workspace to
stomp on.

This way, when tests run after us, they don't have to inherit our cruft.
Yes, we could also have just marked this test as "destructive", but using our
own Test Workspace should be quicker than flushing and recreating the test
environment.