Clean out Users/Workspaces after every test run.
Don't know what I was thinking when I first wrote this... even if we don't
have any Users/Workspaces in place at the start of the test, we should be
cleaning them out at the end.

It _could be_ that the test only needed the "db" fixture and set up a DB, then
went ahead and created some Users/Workspaces for testing.  We might not have
picked up a list of Users/Workspaces when we started, but we *would* have junk
left around that we'd need to clean up after the test run.