Script to help figure out why unit tests fail.
Its not magical, but it tries to help you zero in on the "smallest number of
tests necessary in order to reproduce the failure".

Run the script, pointing it at a branch and a failing unit test.  Using your
checked-out copy of the code and the STCI results off of Galena, this script:

1)  finds the failing test,
2)  verifies that the test passes when run on a clean environment,
3)  assembles a list of likely candidates of tests that ran before it that may
    have screwed up the test environment, and
4)  runs the tests repeatedly until it can zero in on just the smallest set of
    tests that reproduce the failure

This *can* take a few minutes, depending on how many candidate tests get
selected; its not uncommon for it to take 6-7 minutes to zero in on the
smallest set of tests.  That said, if you had to do it manually you could
easily spend that much time poking at it trying to do the same thing.